+++ packages/react-native-nitro-auth/src/global.d.ts
@@ -6,10 +6,11 @@ declare global {
           clientId: string;
           scope: string;
           redirectURI: string;
+          nonce?: string;
           usePopup: boolean;
         }) => void;
         signIn: () => Promise<{
-          authorization: { id_token: string };
+          authorization: { id_token: string; code?: string };
           user?: {
             email?: string;
             name?: {